rename: TableComponents.js to TableComponents.tsx
