+++ benchexec/tablegenerator/react-table/src/components/ReactTable.js
@@ -21,7 +21,7 @@ import {
   StandardCell,
   StandardColumnHeader,
   SelectColumnsButton,
-} from "./TableComponents.js";
+} from "./TableComponents";
 import {
   isNumericColumn,
   numericSortMethod,
